+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Ability to run remote from AD FS server
 ### Changed
 - Fixed naming of internal functions to use approved PowerShell verbs
 - Switched to MIT license